pyproject.toml
[tool.poetry]
name = "yappi23"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "13.12"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pyaspeller = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
